use DET CSFLE setup scripts
