+++ .evergreen/generated_configs/functions.yml
@@ -327,6 +327,47 @@ functions:
       args:
         - -c
         - .evergreen/scripts/compile.sh
+  csfle-setup:
+    - command: subprocess.exec
+      type: setup
+      params:
+        binary: bash
+        working_dir: drivers-evergreen-tools/.evergreen/csfle
+        args:
+          - -c
+          - |
+            set -o errexit
+            echo "Preparing CSFLE venv environment..."
+            # This ensures future invocations of activate-kmstlsvenv.sh conducted in
+            # parallel do not race to setup a venv environment; it has already been prepared.
+            # This primarily addresses the situation where the "test" and "csfle-setup"
+            # functions invoke 'activate-kmstlsvenv.sh' simultaneously.
+            . ./activate-kmstlsvenv.sh && deactivate
+            echo "Preparing CSFLE venv environment... done."
+    - command: ec2.assume_role
+      params:
+        role_arn: ${aws_test_secrets_role}
+    - command: subprocess.exec
+      type: setup
+      params:
+        binary: bash
+        working_dir: drivers-evergreen-tools/.evergreen/csfle
+        include_expansions_in_env:
+          - AWS_ACCESS_KEY_ID
+          - AWS_SECRET_ACCESS_KEY
+          - AWS_SESSION_TOKEN
+        args:
+          - -c
+          - ./setup.sh
+  csfle-teardown:
+    - command: subprocess.exec
+      type: setup
+      params:
+        binary: bash
+        working_dir: drivers-evergreen-tools/.evergreen/csfle
+        args:
+          - -c
+          - ./teardown.sh
   docker-image-build:
     - command: expansions.update
       params:
@@ -433,41 +474,6 @@ functions:
         args:
           - -c
           - .evergreen/scripts/compile-scan-build.sh
-  run_kms_servers:
-    - command: subprocess.exec
-      type: setup
-      params:
-        binary: bash
-        args:
-          - -c
-          - |
-            set -o errexit
-            echo "Preparing CSFLE venv environment..."
-            cd ./drivers-evergreen-tools/.evergreen/csfle
-            # This function ensures future invocations of activate-kmstlsvenv.sh conducted in
-            # parallel do not race to setup a venv environment; it has already been prepared.
-            # This primarily addresses the situation where the "test" and "run_kms_servers"
-            # functions invoke 'activate-kmstlsvenv.sh' simultaneously.
-            . ./activate-kmstlsvenv.sh && deactivate
-            echo "Preparing CSFLE venv environment... done."
-    - command: subprocess.exec
-      type: setup
-      params:
-        binary: bash
-        background: true
-        args:
-          - -c
-          - |
-            set -o errexit
-            echo "Starting mock KMS servers..."
-            cd ./drivers-evergreen-tools/.evergreen/csfle
-            . ./activate-kmstlsvenv.sh
-            python -u kms_http_server.py --ca_file ../x509gen/ca.pem --cert_file ../x509gen/server.pem --port 8999 &
-            python -u kms_http_server.py --ca_file ../x509gen/ca.pem --cert_file ../x509gen/expired.pem --port 9000 &
-            python -u kms_http_server.py --ca_file ../x509gen/ca.pem --cert_file ../x509gen/wrong-host.pem --port 9001 &
-            python -u kms_http_server.py --ca_file ../x509gen/ca.pem --cert_file ../x509gen/server.pem --port 9002 --require_client_cert &
-            python -u kms_kmip_server.py &
-            echo "Starting mock KMS servers... done."
   set-cache-dir:
     - command: subprocess.exec
       type: setup
@@ -533,7 +539,7 @@ functions:
           - |
             set -o errexit
             if test -d drivers-evergreen-tools; then
-                pushd drivers-evergreen-tools
+                cd drivers-evergreen-tools
                 .evergreen/run-mongodb.sh stop
             fi
   test:
@@ -549,13 +555,6 @@ functions:
         - ASAN_SYMBOLIZER_PATH
         - build_type
         - CRYPT_SHARED_LIB_PATH
-        - cse_aws_access_key_id
-        - cse_aws_secret_access_key
-        - cse_azure_client_id
-        - cse_azure_client_secret
-        - cse_azure_tenant_id
-        - cse_gcp_email
-        - cse_gcp_privatekey
         - disable_slow_tests
         - distro_id
         - example_projects_cxxflags
