+++ .evergreen/scripts/test.sh
@@ -4,13 +4,6 @@ set -o errexit
 set -o pipefail
 
 : "${build_type:?}"
-: "${cse_aws_access_key_id:?}"
-: "${cse_aws_secret_access_key:?}"
-: "${cse_azure_client_id:?}"
-: "${cse_azure_client_secret:?}"
-: "${cse_azure_tenant_id:?}"
-: "${cse_gcp_email:?}"
-: "${cse_gcp_privatekey:?}"
 : "${MONGOCXX_TEST_TOPOLOGY:?}"
 
 : "${ASAN_SYMBOLIZER_PATH:-}"
@@ -120,7 +113,8 @@ else
   # export environment variables for encryption tests
   set +o errexit
 
-  echo "Setting temporary credentials..."
+  echo "Setting KMS credentials from the environment..."
+  # Load secrets generated by the "csfle-setup" function (setup.sh -> secrets-export.sh).
   pushd "${DRIVERS_TOOLS:?}/.evergreen/csfle"
   {
     # DEVPROD-4630: use BASH_XTRACEFD instead:
@@ -128,55 +122,28 @@ else
     is_xtrace_set="$([[ "$-" == *x* ]] && echo 1)"
     set +o xtrace
 
-    export AWS_SECRET_ACCESS_KEY="${cse_aws_secret_access_key:?}"
-    export AWS_ACCESS_KEY_ID="${cse_aws_access_key_id:?}"
-    export AWS_DEFAULT_REGION="us-east-1"
-
-    # DEVPROD-4630: use BASH_XTRACEFD instead:
-    #     unset BASH_XTRACEFD
-    [[ "${is_xtrace_set:-}" == 1 ]] && set -o xtrace
-  }
-  echo "Running activate-kmstlsvenv.sh..."
-  # shellcheck source=/dev/null
-  . ./activate-kmstlsvenv.sh
-  echo "Running activate-kmstlsvenv.sh... done."
-  echo "Running set-temp-creds.sh..."
-  # shellcheck source=/dev/null
-  . ./set-temp-creds.sh
-  echo "Running set-temp-creds.sh... done."
-  deactivate
-  popd # "${DRIVERS_TOOLS:?}/.evergreen/csfle"
-  echo "Setting temporary credentials... done."
-
-  # Ensure temporary credentials were properly set.
-  if [ -z "${CSFLE_AWS_TEMP_ACCESS_KEY_ID:-}" ]; then
-    echo "Failed to set temporary credentials!"
-    exit 1
-  fi
-
-  {
-    # DEVPROD-4630: use BASH_XTRACEFD instead:
-    #     exec {BASH_XTRACEFD}>/dev/null
-    is_xtrace_set="$([[ "$-" == *x* ]] && echo 1)"
-    set +o xtrace
+    # shellcheck source=/dev/null
+    . ./secrets-export.sh
 
     export MONGOCXX_TEST_CSFLE_TLS_CA_FILE=${DRIVERS_TOOLS:?}/.evergreen/x509gen/ca.pem
     export MONGOCXX_TEST_CSFLE_TLS_CERTIFICATE_KEY_FILE=${DRIVERS_TOOLS:?}/.evergreen/x509gen/client.pem
-    export MONGOCXX_TEST_AWS_TEMP_ACCESS_KEY_ID="$CSFLE_AWS_TEMP_ACCESS_KEY_ID"
-    export MONGOCXX_TEST_AWS_TEMP_SECRET_ACCESS_KEY="$CSFLE_AWS_TEMP_SECRET_ACCESS_KEY"
-    export MONGOCXX_TEST_AWS_TEMP_SESSION_TOKEN="$CSFLE_AWS_TEMP_SESSION_TOKEN"
-    export MONGOCXX_TEST_AWS_SECRET_ACCESS_KEY="${cse_aws_secret_access_key:?}"
-    export MONGOCXX_TEST_AWS_ACCESS_KEY_ID="${cse_aws_access_key_id:?}"
-    export MONGOCXX_TEST_AZURE_TENANT_ID="${cse_azure_tenant_id:?}"
-    export MONGOCXX_TEST_AZURE_CLIENT_ID="${cse_azure_client_id:?}"
-    export MONGOCXX_TEST_AZURE_CLIENT_SECRET="${cse_azure_client_secret:?}"
-    export MONGOCXX_TEST_GCP_EMAIL="${cse_gcp_email:?}"
-    export MONGOCXX_TEST_GCP_PRIVATEKEY="${cse_gcp_privatekey:?}"
+    export MONGOCXX_TEST_AWS_TEMP_ACCESS_KEY_ID="${CSFLE_AWS_TEMP_ACCESS_KEY_ID:?}"
+    export MONGOCXX_TEST_AWS_TEMP_SECRET_ACCESS_KEY="${CSFLE_AWS_TEMP_SECRET_ACCESS_KEY:?}"
+    export MONGOCXX_TEST_AWS_TEMP_SESSION_TOKEN="${CSFLE_AWS_TEMP_SESSION_TOKEN:?}"
+    export MONGOCXX_TEST_AWS_SECRET_ACCESS_KEY="${FLE_AWS_SECRET:?}"
+    export MONGOCXX_TEST_AWS_ACCESS_KEY_ID="${FLE_AWS_KEY:?}"
+    export MONGOCXX_TEST_AZURE_TENANT_ID="${FLE_AZURE_TENANTID:?}"
+    export MONGOCXX_TEST_AZURE_CLIENT_ID="${FLE_AZURE_CLIENTID:?}"
+    export MONGOCXX_TEST_AZURE_CLIENT_SECRET="${FLE_AZURE_CLIENTSECRET:?}"
+    export MONGOCXX_TEST_GCP_EMAIL="${FLE_GCP_EMAIL:?}"
+    export MONGOCXX_TEST_GCP_PRIVATEKEY="${FLE_GCP_PRIVATEKEY:?}"
 
     # DEVPROD-4630: use BASH_XTRACEFD instead:
     #     unset BASH_XTRACEFD
     [[ "${is_xtrace_set:-}" == 1 ]] && set -o xtrace
   }
+  popd # "${DRIVERS_TOOLS:?}/.evergreen/csfle"
+  echo "Setting KMS credentials from the environment... done."
 
   set -o errexit
 
@@ -224,7 +191,6 @@ else
     return 1
   }
   echo "Waiting for mock KMS servers to start..."
-  wait_for_kms_server 8999
   wait_for_kms_server 9000
   wait_for_kms_server 9001
   wait_for_kms_server 9002
